[APIS-919] 11.2 Driver Conformance Test - ADO.NET  (#34)
http://jira.cubrid.org/browse/APIS-919

- update version (11.1.0.0001)
- update CCI dirver (11.1.0)
- add script for get cci driver library
- modify ado.net driver for count type (int -> bigint)
- added charset in 'T_CCI_COL_INFO ' struct  (sync to CCI Driver)
    (issue : convert error in reader(string -> string[], null error in DataAdapter.Fill)
- update testcase for count type (int -> bigint)
- update testcase for user schema
- update testcase for oid value (11.2 - game oid is '@4481|1|0')
- modify testcase due to default value(DONT_REUSE_OID) change in server 11.2
- modify the wrong testcase (GetSchema(Views).Rows.Count Value)
- added create function(Procedure)  code for testcase (GetSchema_NameAndFilter_Test)

+++ Code/Src/CUBRIDDataReader.cs
@@ -44,7 +44,7 @@ namespace CUBRID.Data.CUBRIDClient
   public sealed class CUBRIDDataReader : DbDataReader, IDataRecord
   {
     private  ColumnMetaData[] columnMetaData;
-    private  int resultCount; //Resultset rows count; Equivalent to SELECT COUNT(*)...
+    private  Int64 resultCount; //Resultset rows count; Equivalent to SELECT COUNT(*)...
     private  ResultTuple resultTuple;
     private CommandBehavior CommandBehavior; //TODO Add support
     private CUBRIDConnection conn;
