+++ Code/Test/Functional/TestCUBRIDConnection.cs
@@ -405,7 +405,7 @@ namespace Test.Functional
                 command.ExecuteNonQuery();
             }
 
-            int tablesCount = GetTablesCount("t", conn);
+            Int64 tablesCount = GetTablesCount("t", conn);
             Debug.Assert(tablesCount == 1);
 
             conn.Rollback();
